+++ app/src/main/java/com/example/cinevote/data/database/Room/FIlm.kt
@@ -3,14 +3,13 @@ package com.example.cinevote.data.database.Room
 import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.PrimaryKey
-import java.util.Date
 
 @Entity
 data class FilmList(
-    @PrimaryKey val title : String,
-    @ColumnInfo val posterPath : String,
+    @PrimaryKey val title: String,
+    @ColumnInfo val posterPath: String,
     @ColumnInfo val plot: String,
-    @ColumnInfo val voteAverage: Int,
+    @ColumnInfo val voteAverage: Float,
     @ColumnInfo val releaseDate: String,
-    @ColumnInfo val genreIDs : String,
+    @ColumnInfo val genreIDs: String,
 )
